ai_bot/requirements.txt
google-generativeai
pillow
requests
schedule